fix polynom
DIMENSION の型は int ではなく size_t です

+++ polynom.c
@@ -1,6 +1,6 @@
 #include "polynom.h"
 
-int DIMENSION = 4096;
+size_t DIMENSION = 4096;
 
 int InitPolynom(polynom *pp)
 {
